# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart LPR API",
        "description": "Open API for Smart LPR",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/accounts/{page}/{limit}": {
            "get": {
                "tags": [
                    "accounts"
                ],
                "summary": "Get accounts of items",
                "description": "Retrieve a accounts of items based on date range, gate, pagination, and limit",
                "operationId": "accounts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Limit number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List retrieved successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/accounts/{id}": {
            "get": {
                "tags": [
                    "accounts"
                ],
                "summary": "Get account by ID",
                "operationId": "92d5e52983a42663a671eadf91d4e2a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of account to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "account not found"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Login",
                "operationId": "34d49de46e0fc2b1c6af3db9a4e3d7f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Username",
                                        "type": "string",
                                        "example": "username"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}